refactore button components
