+++ src/components/navbar.jsx
@@ -3,8 +3,8 @@ import { useDarkMode } from '../hooks/useDarkMode'
 import { useNavMode } from '../hooks/useNavMode'
 import useScrollPosition from '../hooks/useScrollPosition'
 import BannerCarousel from './bannerCarousel'
-import ButtonIcon from './buttonIcon'
-import LinkButton from './linkButton'
+import ButtonIcon from './buttons/buttonIcon'
+import LinkButton from './buttons/linkButton'
 import { selectUserEmail } from '../redux/features/userSlice'
 import { useSelector, useDispatch } from 'react-redux'
 import { useEffect, useState } from 'react'
