--- src/components/purchaseItem.jsx
@@ -1,5 +1,0 @@
-function PurchaseItem() {
-  return <div>This is purchase Item</div>
-}
-
-export default PurchaseItem
