--- src/components/purchaseList.jsx
@@ -1,5 +1,0 @@
-function PurchaseList() {
-  return <div>This is purchase List</div>
-}
-
-export default PurchaseList
