+++ src/pages/auth/forgetPassword.jsx
@@ -3,8 +3,9 @@ import { useForm } from 'react-hook-form'
 import { motion } from 'framer-motion'
 import { Helmet } from 'react-helmet-async'
 import { yupResolver } from '@hookform/resolvers/yup'
+
 import { forgetPasswordScheme } from '../../validations/forgetPassword'
-import Button from '../../components/button'
+import { Button } from '../../components/buttons'
 import { resetPasswordByEmail } from '../../services/auth'
 import WhiteRoomImg from '../../assets/images/features3.png'
