+++ src/pages/checkout/information.jsx
@@ -1,6 +1,9 @@
 import { yupResolver } from '@hookform/resolvers/yup'
 import { useForm } from 'react-hook-form'
 import { useSelector, useDispatch } from 'react-redux'
+import { useEffect, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
 import {
   selectUserAddressDefault,
   selectUserEmail,
@@ -18,9 +21,7 @@ import Controller from '../../components/popup/controller'
 import Input from '../../components/popup/input'
 import EmailSchema from '../../validations/email'
 import AddressList from '../../components/user/addressList'
-import { useEffect, useState } from 'react'
-import Button from '../../components/button'
-import { useNavigate } from 'react-router-dom'
+import { Button } from '../../components/buttons'
 import PopupAddress from '../../components/popup/popupAddress'
 
 function CheckoutInformation() {
