+++ src/pages/products/category.jsx
@@ -6,7 +6,7 @@ import { useParams } from 'react-router-dom'
 import clubImg from '../../assets/images/features3.png'
 import EmailField from '../../components/emailField'
 import ProductItemListing from '../../components/productItemListing'
-import Button from '../../components/button'
+import { Button } from '../../components/buttons'
 import {
   selectProductsCategory,
   getCategoryProducts,
