+++ src/pages/products/productCart.jsx
@@ -1,8 +1,9 @@
 import { Helmet } from 'react-helmet-async'
 import { Link } from 'react-router-dom'
 import { useSelector } from 'react-redux'
+
 import ProductCartList from '../../components/productCartList'
-import Button from '../../components/button'
+import { Button } from '../../components/buttons'
 import { selectCartItems } from '../../redux/features/carts/cartSlice'
 import { selectCurrentStep } from '../../redux/features/stepper/stepperSlice'
