+++ src/pages/products/productListing.jsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Helmet } from 'react-helmet-async'
-import Button from '../../components/button'
+
+import { Button } from '../../components/buttons'
 import clubImg from '../../assets/images/features3.png'
 import EmailField from '../../components/emailField'
 import ProductItemListing from '../../components/productItemListing'
